Changing to pytest and introducing nox for running tests.
