+++ .coveragerc
@@ -1,4 +1,5 @@
 [report]
+show_missing = True
 exclude_lines =
     pragma: no cover
     def __repr__
