+++ setup.cfg
@@ -1,4 +1,2 @@
-[nosetests]
-nologcapture=1 
-with-coverage=1
-cover-package=gpflowopt
+[bdist_wheel]
+universal = 1
