+++ testing/unit/test_acquisition.py
@@ -3,7 +3,7 @@ import numpy as np
 import gpflow
 import tensorflow as tf
 from parameterized import parameterized
-from .utility import create_parabola_model, parabola2d, plane, GPflowOptTestCase
+from ..utility import create_parabola_model, parabola2d, plane, GPflowOptTestCase
 
 domain = np.sum([gpflowopt.domain.ContinuousParameter("x{0}".format(i), -1, 1) for i in range(1, 3)])
 
@@ -326,25 +326,4 @@ class TestJointAcquisition(GPflowOptTestCase):
             self.assertListEqual(joint.operands.sorted_params, [acq1, acq2, acq3, acq4])
 
 
-class TestRecompile(GPflowOptTestCase):
-    """
-    Regression test for #37
-    """
-    def test_vgp(self):
-        with self.test_session():
-            domain = gpflowopt.domain.UnitCube(2)
-            X = gpflowopt.design.RandomDesign(10, domain).generate()
-            Y = np.sin(X[:,[0]])
-            m = gpflow.vgp.VGP(X, Y, gpflow.kernels.RBF(2), gpflow.likelihoods.Gaussian())
-            acq = gpflowopt.acquisition.ExpectedImprovement(m)
-            m.compile()
-            self.assertFalse(m._needs_recompile)
-            acq.evaluate(gpflowopt.design.RandomDesign(10, domain).generate())
-            self.assertTrue(hasattr(acq, '_evaluate_AF_storage'))
 
-            Xnew = gpflowopt.design.RandomDesign(5, domain).generate()
-            Ynew = np.sin(Xnew[:,[0]])
-            acq.set_data(np.vstack((X, Xnew)), np.vstack((Y, Ynew)))
-            self.assertFalse(hasattr(acq, '_needs_recompile'))
-            self.assertFalse(hasattr(acq, '_evaluate_AF_storage'))
-            acq.evaluate(gpflowopt.design.RandomDesign(10, domain).generate())
