+++ testing/unit/test_design.py
@@ -1,7 +1,7 @@
 import gpflowopt
 import numpy as np
 import os
-from .utility import GPflowOptTestCase
+from ..utility import GPflowOptTestCase
 
 
 class _TestDesign(object):
@@ -74,7 +74,7 @@ class TestLatinHyperCubeDesign(_TestDesign, GPflowOptTestCase):
         return [gpflowopt.design.LatinHyperCube(20, domain) for domain in self.domains]
 
     def test_validity(self):
-        groundtruth = np.load(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data', 'lhd.npz'))
+        groundtruth = np.load(os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'data', 'lhd.npz'))
         points = [lhd.generate() for lhd in self.designs]
         lhds = map(lambda file: groundtruth[file], groundtruth.files)
         idx = np.argsort([lhd.shape[-1] for lhd in lhds])
